+++ packages/schema/bind/src/bindings/assemblyscript/wasm-as/baseTypes.ts
