+++ packages/schema/bind/src/bindings/index.ts
@@ -1,18 +1,12 @@
-import { OutputDirectory, BindLanguage } from "../";
+import { GenerateBindingFn } from "./types";
 import * as WasmAs from "./assemblyscript/wasm-as";
-import * as PluginTs from "./typescript/plugin-ts";
-import * as AppTs from "./typescript/app-ts";
+/*import * as PluginTs from "./typescript/plugin-ts";
+import * as AppTs from "./typescript/app-ts";*/
 
-import { TypeInfo } from "@web3api/schema-parse";
+import { BindLanguage } from "../";
 
-export { WasmAs, PluginTs, AppTs };
-
-export type GenerateBindingFn = (
-  output: OutputDirectory,
-  typeInfo: TypeInfo,
-  schema: string,
-  config: Record<string, unknown>
-) => void;
+export { WasmAs/*, PluginTs, AppTs*/ };
+export * from "./types";
 
 export function getGenerateBindingFn(
   bindLanguage: BindLanguage
@@ -20,26 +14,11 @@ export function getGenerateBindingFn(
   switch (bindLanguage) {
     case "wasm-as":
       return WasmAs.generateBinding;
-    case "plugin-ts":
+    /*case "plugin-ts":
       return PluginTs.generateBinding;
     case "app-ts":
-      return AppTs.generateBinding;
+      return AppTs.generateBinding;*/
     default:
       throw Error(`Error: Language binding unsupported - ${bindLanguage}`);
   }
-}
-
-export function generateBinding(
-  bindLanguage: BindLanguage,
-  typeInfo: TypeInfo,
-  schema: string,
-  config: Record<string, unknown>
-): OutputDirectory {
-  const output: OutputDirectory = {
-    entries: [],
-  };
-
-  getGenerateBindingFn(bindLanguage)(output, typeInfo, schema, config);
-
-  return output;
 }
