+++ packages/schema/bind/src/bindings/types.ts
@@ -1,4 +1,10 @@
-export type MustacheFunction = () => (
+import { BindOptions, BindOutput } from "../";
+
+export type GenerateBindingFn = (
+  options: BindOptions
+) => BindOutput;
+
+export type MustacheFn = () => (
   value: string,
   render: (template: string) => string
 ) => string;
