+++ packages/schema/bind/src/bindings/typescript/functions.ts
@@ -1,6 +1,6 @@
-import { MustacheFunction } from "../types";
+import { MustacheFn } from "../types";
 
-export const toTypescript: MustacheFunction = () => {
+export const toTypescript: MustacheFn = () => {
   return (value: string, render: (template: string) => string) => {
     let type = render(value);
