+++ packages/schema/bind/src/bindings/utils/path.ts
