--- packages/schema/bind/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from "./fs";
-export * from "./path";
-export * from "./typeInfo";
