+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/AnotherType/serialization.ts
