+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/CustomType/serialization.ts
