+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/imported/TestImport_AnotherObject/serialization.ts
