+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/imported/TestImport_Object/serialization.ts
