+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/index.ts
@@ -1,0 +1,17 @@
+export { CustomType } from "./CustomType";
+export { AnotherType } from "./AnotherType";
+export {
+  CustomEnum,
+  getCustomEnumKey,
+  getCustomEnumValue,
+  sanitizeCustomEnumValue
+} from "./CustomEnum";
+export { TestImport_Query } from "./imported/TestImport_Query";
+export { TestImport_Object } from "./imported/TestImport_Object";
+export { TestImport_AnotherObject } from "./imported/TestImport_AnotherObject";
+export {
+  TestImport_Enum,
+  getTestImport_EnumKey,
+  getTestImport_EnumValue,
+  sanitizeTestImport_EnumValue
+} from "./imported/TestImport_Enum";
