+++ packages/test-cases/cases/bind/sanity/output/wasm-as/mutation/common.ts
@@ -1,1 +1,1 @@
-export * from "../query/common";
+export * from "../common/common";
