+++ packages/test-cases/cases/bind/sanity/output/wasm-as/query/common.ts
@@ -1,8 +1,1 @@
-export * from "./CustomType";
-export * from "./AnotherType";
-export * from "./CustomEnum";
-
-export * from "./imported/TestImport_Query";
-export * from "./imported/TestImport_Object";
-export * from "./imported/TestImport_AnotherObject";
-export * from "./imported/TestImport_Enum";
+export * from "../common/common";
